코드 정리(Location1 삭제) 현위치 및 주변 문화 정보 업데이트 따로 메소드로 분리
